add tuto image, about page, device field report

+++ Mobile/Android/Barde/app/src/main/java/com/project/barde/barde/ui/view/SlideBarListItemView.kt
@@ -33,9 +33,9 @@ class SlideBarListItemView @JvmOverloads constructor(
         mValue = slideModel.min.toInt()
         m = (mMax - mMin) / 100.0f
         slide_bar_list_item_view_title.text = slideModel.title
-        slide_bar_list_item_view_min.text = slideModel.min.toString()
-        slide_bar_list_item_view_max.text = slideModel.max.toString()
-        slide_bar_list_item_view_value.text = mValue.toString()
+        slide_bar_list_item_view_min.text = slideModel.min.toInt().toString()
+        slide_bar_list_item_view_max.text = slideModel.max.toInt().toString()
+        slide_bar_list_item_view_value.text = mValue.toInt().toString()
         slide_bar_list_item_view_unit.text = slideModel.unit
         slide_bar_list_item_view_seek_bar.setOnSeekBarChangeListener(object : SeekBar.OnSeekBarChangeListener {
